polish code
