+++ python/paddle/fluid/tests/unittests/test_softmax_op.py
@@ -112,73 +112,9 @@ class TestSoftmaxOp(OpTest):
             )
 
 
-class TestSoftmaxOpfp32(OpTest):
-    def get_x_shape(self):
-        return [10, 10]
-
-    def get_axis(self):
-        return -1
-
-    def setUp(self):
-        self.op_type = "softmax"
-        self.comp_op_type = "comp"
-        self.python_api = F.softmax
-        self.use_cudnn = False
-        self.use_mkldnn = False
-        # explicilty use float32 for ROCm, as MIOpen does not yet support float64
-        self.dtype = np.float32
-        self.init_kernel_type()
-        self.shape = self.get_x_shape()
-        self.axis = self.get_axis()
-
-        np.random.seed(0)
-        x = np.random.uniform(0.1, 1, self.shape).astype(self.dtype)
-        out = np.apply_along_axis(stable_softmax, self.axis, x)
-
-        self.inputs = {'X': OpTest.np_dtype_to_fluid_dtype(x)}
-        self.outputs = {'Out': out}
-        self.attrs = {
-            'axis': self.axis,
-            'use_cudnn': self.use_cudnn,
-            'use_mkldnn': self.use_mkldnn,
-        }
-        # self.enable_cinn = False
-
+class TestSoftmaxOpfp32(TestSoftmaxOp):
     def init_kernel_type(self):
-        pass
-
-    def test_check_output(self):
-        # TODO(wangzhongpu): support mkldnn op in dygraph mode
-        if self.use_cudnn:
-            place = core.CUDAPlace(0)
-            self.check_output_with_place(
-                place, atol=1e-5, check_dygraph=(not self.use_mkldnn)
-            )
-        else:
-            self.check_output(
-                check_dygraph=(not self.use_mkldnn), check_comp=True
-            )
-
-    def test_check_grad(self):
-        # TODO(wangzhongpu): support mkldnn op in dygraph mode
-        if self.use_cudnn or self.dtype == np.float16:
-            place = core.CUDAPlace(0)
-            if core.is_float16_supported(place):
-                self.check_grad_with_place(
-                    place,
-                    ["X"],
-                    "Out",
-                    max_relative_error=0.01,
-                    check_dygraph=(not self.use_mkldnn),
-                )
-        else:
-            self.check_grad(
-                ["X"],
-                "Out",
-                max_relative_error=0.01,
-                check_dygraph=(not self.use_mkldnn),
-                check_comp=True,
-            )
+        self.dtype = np.float32
 
 
 class TestSoftmaxOp_ZeroDim1(TestSoftmaxOp):
@@ -214,27 +150,6 @@ class TestSoftmaxOp_ZeroDim1(TestSoftmaxOp):
         else:
             self.check_output(
                 check_dygraph=(not self.use_mkldnn), check_comp=True
-            )
-
-    def test_check_grad(self):
-        # TODO(wangzhongpu): support mkldnn op in dygraph mode
-        if self.use_cudnn or self.dtype == np.float16:
-            place = core.CUDAPlace(0)
-            if core.is_float16_supported(place):
-                self.check_grad_with_place(
-                    place,
-                    ["X"],
-                    "Out",
-                    max_relative_error=0.01,
-                    check_dygraph=(not self.use_mkldnn),
-                )
-        else:
-            self.check_grad(
-                ["X"],
-                "Out",
-                max_relative_error=0.01,
-                check_dygraph=(not self.use_mkldnn),
-                check_comp=True,
             )
 
 
@@ -273,27 +188,6 @@ class TestSoftmaxOp_ZeroDim2(TestSoftmaxOp):
         else:
             self.check_output(
                 check_dygraph=(not self.use_mkldnn), check_comp=True
-            )
-
-    def test_check_grad(self):
-        # TODO(wangzhongpu): support mkldnn op in dygraph mode
-        if self.use_cudnn or self.dtype == np.float16:
-            place = core.CUDAPlace(0)
-            if core.is_float16_supported(place):
-                self.check_grad_with_place(
-                    place,
-                    ["X"],
-                    "Out",
-                    max_relative_error=0.01,
-                    check_dygraph=(not self.use_mkldnn),
-                )
-        else:
-            self.check_grad(
-                ["X"],
-                "Out",
-                max_relative_error=0.01,
-                check_dygraph=(not self.use_mkldnn),
-                check_comp=True,
             )
