+++ python/paddle/jit/dy2static/program_translator.py
@@ -1222,6 +1222,9 @@ class ProgramCache:
     def concrete_programs(self):
         return [cp for key, (cp, _) in self._caches.items()]
 
+    def clear(self):
+        self._caches = collections.OrderedDict()
+
 
 class ProgramTranslator:
     """
